Glyco score overhaul (#20)
* support >1 PSM per MS2 spectrum (DDA+/DIA)

* add test for >1 result per PSM file editing

* update gitignore

* add option to skip Y fragment scores on 2nd pass

to score using spectral similarity only

* add general oxonium ion spectrum similarity score

* update oxonium ion default list

* include spec similarity scores in glycan pairwise comparisons

* correct decoy glycan mass when matching

so that decoy and target versions of a glycan are always considered for a given spectrum. (i.e., decoy mass shift only affects scores)

* do not perform LDA on 1st pass

* remove glycans not present in 1st pass library from database for 2nd pass

* set default min iso prob

fixes a crash introduced by always matching decoy type 1, which may have isotopes other than those set by user

* add option to test normed mass score

* remove mass2nd option

found to be very useful - always include

* add centroid MS1 score

* remove deprecated yprop score

* remove deprecated iso score

* remove param, always normalized Y and oxo fragment intensities

* remove testing param for sim-only 2nd pass

always use similiarity only on 2nd pass

* rename 'static' and 'dynamic' scores to 1st and 2nd pass and remove deprecated score types

some scores unreachable now that 2nd pass is similarity only

* remove deprecated param glycoYnorm

* remove param norm mass score and deprecated mass and isotope scoring methods

* remove deprecated noncomp FDR param

noncompetitive FDR always used as a fallback, but never the initial method

* remove deprecated default glyco propensity param

never used, and trivial now that propensity is only for pairwise diagnostic oxos on 2nd pass

* reorganize and rename glycan score methods

* fix normed mass score

* fix similarity score NaNs and extreme values

* fix normalized mass error score

* replace empirical mass score with norm mass score in 1st pass

* remove deprecated lda feature params and clean up rawglyco output

always use Y, ox (diagnostic), ox (general), and mass scores

* fix not using diagnostic oxo scores in 2nd pass absolute score

* fix entropy score

* add option to remove glycans in 2nd pass

* add cosine sim score option

* upgrade batmass-io to 1.36.7

* bump to 3.0.12

* fix glyco decoy fragment init for shuffled decoys

* remove empirical diagnostic oxo score from 2nd pass

* bump to 3.0.13

* add average decoy type and decoy_fragment_type param

* add nearest glycan decoy fragment method

* add nearest glycan decoy method

* add decoy fragment types 4,5,6 for testing

* remove deprecated fragment propensity info

* make decoys from low quality targets, add options for avg vs median

* add back specific oxo score to 2nd pass, but as similarity rather than empirical

* change to use alternate (target) composition matches as decoys

* change hashmaps to linkedhashmaps to avoid inconsistency on reruns due to key retrieval order

# Conflicts:
#	src/edu/umich/andykong/ptmshepherd/glyco/GlycoAnalysis.java

* fix filtering of glycans with no mass match after database reduction

* change to use all alternate matched glycans for decoy spectra instead of first one

* change mod from float to double to prevent loss of precision for some glyco operations

* change psm delta mass to double

* do not use low confidence matches for glyco decoys

they can be misassigned target spectra

* change to use random selection of high confidence glyco PSMs for decoy generation fallback

* performance improvement for glyco decoy generation

* implement multipass glycan assign

* option to use random for all decoys

instead of nearby targets first, then random

* fix a test

* remove deprecated options for decoy fragment generation

removed low score targets map, and high score targets -> decoy method

* add multi-decoy option for glycan assignment

* add variable mod check module for glycan assignment

* fix randomness on repeated runs

* refactor to search spectra for list of ions instead of single ion (glyco)

* refactor diagnostic mining to search spectra for all ions simultaneously

* fix MS1 score and add MS1 delta score

* refactor glyco fragment matching to a method

* fix MS1 spectra file type priorities

* add no-FDR mode for glyco analysis

* additional fixes for passing .d files to IonQuant API for glyco

* pre-build ionquant indices for glyco

if there aren't too many raw files

* print general ox to rawglyco file

* minor cleanup

* add glycan library option for similarity-based 1st pass

* add min fragment diff param for decoys

* add option to skip pairwise for glyco

* add update glycolib option

* decouple glycan decoy shifts from input ppm and isotope tolerances

* add option to 5x 1st pass FDR

* add NN scoring option

* run LDA/NN on 1st pass if using library

* fix single pass if not using two pass mode

* remove deprecated testing params

cosine sim, avg ints, and increase 1st fdr

* clean up glyco lib and skip pairwise params

always skip pairwise on "lib" runs (always on 2nd pass, and on 1st pass if using external lib)

* add optional glyco score distribution plots

# Conflicts:
#	src/edu/umich/andykong/ptmshepherd/PTMShepherd.java
#	src/edu/umich/andykong/ptmshepherd/glyco/GlycoParams.java

* update parameter defaults for glyco

* update glyco param printing

* move glycan plot outputs to subfolder

* change xchart version to match fragpipe

* fix skipMS1 scoring param for glyco

sync with fragpipe

* package isotope dependency even in package no deps

because it's not packaged in fragpipe (internal to ionquant)

* upgrade batmass-io 1.36.11

* prevent AWT from keeping JVM alive after program exits

* allow glyco module to run without ionquant API if it is not needed

needed to make it still work for non-IonQuant users

* fix calling for ms1 features if skipMS1 is true (glyco)

* fix non-library 2pass mode

* only write rawglyco2 in debug mode

* clean up prev rawglyco files correctly, do not delete final rawglyco file after analysis

* remove some deprecated glyco debug plots

* fix debug plot names for 2nd pass

* write reduced glyco DB for ionquant

* propagate random generator correctly to NN

* increase decimal places for glyco scores

* Update src/edu/umich/andykong/ptmshepherd/Mod.java


* Update src/edu/umich/andykong/ptmshepherd/glyco/GlycanAssignmentResult.java

only print general oxoniums with found int > 0


* fix copilot commit

---------

+++ src/edu/umich/andykong/ptmshepherd/diagnosticmining/DiagnosticAnalysis.java
@@ -147,8 +147,8 @@ public class DiagnosticAnalysis {
         String specName = psm.getSpec();
         int charge = psm.getCharge();
         String pepSeq = psm.getPeptide();
-        float dmass = psm.getDMass();
-        float pepMass = psm.getCalcPepmass();
+        float dmass = (float) psm.getDMass();
+        float pepMass = (float) (double) psm.getCalcPepmass();
         ArrayList<Mod> smods = psm.getAssignedMods();
 
         /* Prep spec and normalize to base peak */
